refactor(employee): consolidate and enhance JS scripts for employee management
Removed Rollup configuration and related dependencies, simplifying the build process by replacing Rollup with a basic echo command in package.json. Consolidated the phone number formatting script into the main employee script, enhancing maintainability and eliminating redundant script references in the employee.html template.

BREAKING CHANGE: Removed Rollup-based build system and related dependencies. Projects relying on Rollup for module bundling need to migrate to alternative solutions or update their build scripts.
