+++ package.json
@@ -1,20 +1,12 @@
 {
   "name": "js-demo",
   "version": "0.0.1",
-  "main": "",
   "type": "module",
   "dependencies": {
     "@popperjs/core": "^2.11.8",
-    "bootstrap": "^5.3.3",
-    "jquery": "^3.7.1"
+    "bootstrap": "^5.3.3"
   },
   "scripts": {
-    "build": "rollup -c"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^26.0.1",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-terser": "^0.4.4",
-    "rollup": "^4.21.1"
+    "build": "echo 'Building...'"
   }
 }
