+++ src/main/resources/templates/employee/employee.html
@@ -4,7 +4,6 @@
     <title>Colaboradores</title>
     <link rel="stylesheet" th:href="@{protected/css/employee-styles.css}">
     <script th:src="@{protected/js/employee/employee-script.js}" defer></script>
-    <script th:src="@{protected/js/employee/formatPhone-script.js}" defer></script>
 </head>
 <body>
 <main layout:fragment="content">
